sbi: move reset extension to its own file
Continue to move extensions to their own files to reduce clutter in
sbi.rs.
